api_requirements.txt
Flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.1.1
flask-wtf==1.1.1
flask-cors==4.0.0
email-validator==2.0.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
requests==2.31.0
stripe==7.2.0
werkzeug==2.3.7
wtforms==3.0.1
sqlalchemy==2.0.21
pyjwt==2.8.0